# server/openapi.yaml
openapi: '3.0.3'
info:
  title: ESA Dono API
  description: |
    Donation platform with Stripe checkout, wallet balance, pledge (smart cart)
    auto-fulfillment, rewards, polls, fund goals, and moderator/admin tooling.

    ## Authentication

    | Method | Mechanism | Used By |
    |--------|-----------|---------|
    | Donor token | `?token=<magic_token>` query param on every request | Donor wallet, spend endpoints, moderator |
    | Admin key | `X-Admin-Key: <ADMIN_API_KEY>` header | Admin routes; also a fallback for moderator routes |
    | Moderator key | `X-Moderator-Key: <MODERATOR_API_KEY>` header | Fallback access to moderator routes, independent of donor roles |
    | Moderator/Admin role | Donor token + `role: MODERATOR` or `role: ADMIN` on donor record | Moderator routes (ADMIN implies MODERATOR access) |
    | Stripe webhook | HMAC signature via `stripe-signature` header | Stripe webhook (skipped when `STRIPE_WEBHOOK_SECRET` unset) |

    ## Rate Limiting

    Spend endpoints (`claim`, `vote`, `custom-entry`, `contribute`) are rate-limited:
    20 requests per minute, keyed by `?token=` or IP. Configurable via `RATE_LIMIT_SPEND`.

    ## Wallet Discount

    When `POST /api/pledge` receives a valid `?token=` (authenticated donor),
    the donor's `balance_remaining` is applied as a discount on the Stripe
    checkout amount. If the wallet covers the entire pledge total, the pledge
...

    ## Money

...
  version: '1.0.0'
servers:
  - url: /api
...
paths:
...
  /moderator/stats:
    get:
      tags: [Moderator]
      summary: Moderator dashboard stats
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Moderator stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending_entries: { type: integer }
                  active_polls: { type: integer }
                  total_rewards: { type: integer }
                  total_goals: { type: integer }
  /moderator/streams:
    get:
      tags: [Moderator]
      summary: List all streams
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: All streams ordered by created_at ascending
    post:
      tags: [Moderator]
      summary: Create a stream
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                is_active: { type: boolean, default: true }
      responses:
        '200':
          description: Stream created
        '409':
          description: Stream name already exists
  /moderator/streams/{id}:
    put:
      tags: [Moderator]
      summary: Update a stream
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                is_active: { type: boolean }
      responses:
        '200':
          description: Stream updated
    delete:
      tags: [Moderator]
      summary: Deactivate a stream
      description: Soft-delete — sets is_active false rather than removing the row, to preserve references from incentives/donations/pledges.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Stream deactivated
  /moderator/polls:
    get:
      tags: [Moderator]
      summary: List all polls
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Polls with options and custom entries
    post:
      tags: [Moderator]
      summary: Create a poll
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                is_active: { type: boolean, default: true }
                ends_at: { type: string, format: date-time, nullable: true }
                options:
                  type: array
                  items:
                    type: object
                    properties:
                      label: { type: string }
                allow_custom_entries: { type: boolean, default: false }
                max_entry_chars: { type: integer, nullable: true }
                auto_approve: { type: boolean, default: true }
                stream_id: { type: string, nullable: true }
      responses:
        '200':
          description: Poll created
  /moderator/polls/{id}:
    put:
      tags: [Moderator]
      summary: Update a poll
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                is_active: { type: boolean }
                ends_at: { type: string, format: date-time, nullable: true }
                allow_custom_entries: { type: boolean }
                max_entry_chars: { type: integer, nullable: true }
                auto_approve: { type: boolean }
                stream_id: { type: string, nullable: true }
      responses:
        '200':
          description: Poll updated
    delete:
      tags: [Moderator]
      summary: Delete a poll
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /moderator/polls/{id}/options:
    post:
      tags: [Moderator]
      summary: Add a poll option
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
      responses:
        '200':
          description: Option created
  /moderator/polls/options/{id}:
    delete:
      tags: [Moderator]
      summary: Delete a poll option
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /moderator/polls/{id}/custom-entries:
    get:
      tags: [Moderator]
      summary: List custom entries for a poll
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Custom entries with option details (no donor email)
  /moderator/polls/custom-entries/{id}:
    patch:
      tags: [Moderator]
      summary: Approve or reject a custom poll entry
      description: |
        APPROVED: credits option tally (`votes_cents`, `total_votes_cents`).
        REJECTED: refunds donor balance, marks option REJECTED, writes BalanceAdjustment.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [APPROVED, REJECTED]
      responses:
        '200':
          description: Entry updated
  /moderator/rewards:
    get:
      tags: [Moderator]
      summary: List all rewards
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Rewards ordered by created_at descending
    post:
      tags: [Moderator]
      summary: Create a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, cost_cents]
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean, default: true }
                custom_type_label: { type: string }
                stream_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward created
  /moderator/rewards/{id}:
    put:
      tags: [Moderator]
      summary: Update a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean }
                custom_type_label: { type: string }
                stream_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward updated
    delete:
      tags: [Moderator]
      summary: Delete a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /moderator/claims:
    get:
      tags: [Moderator]
      summary: List all reward claims
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Claims with reward details (no donor email)
  /moderator/claims/{id}:
    patch:
      tags: [Moderator]
      summary: Update claim status
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [PENDING, FULFILLED]
      responses:
        '200':
          description: Claim updated
  /moderator/donations:
    get:
      tags: [Moderator]
      summary: List all donations (regardless of donor), for moderation review
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Donations with donor name (no donor email) and moderation status
  /moderator/donations/{id}:
    patch:
      tags: [Moderator]
      summary: Mark (or unmark) a donation as moderated
      description: |
        Sets `moderated`, `moderated_at`, and `moderated_by` so downstream
        tools (exports, leaderboards, Discord role sync, etc.) can rely on
        which donations a human has reviewed.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moderated]
              properties:
                moderated: { type: boolean }
      responses:
        '200':
          description: Donation updated
        '400':
          description: moderated must be a boolean
  /moderator/goals:
    get:
      tags: [Moderator]
      summary: List all fund goals
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Goals ordered by created_at descending
    post:
      tags: [Moderator]
      summary: Create a fund goal
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, target_cents]
              properties:
                title: { type: string }
                description: { type: string }
                target_cents: { type: integer }
                stream_id: { type: string, nullable: true }
                is_active: { type: boolean, default: true }
      responses:
        '200':
          description: Goal created
  /moderator/goals/{id}:
    put:
      tags: [Moderator]
      summary: Update a fund goal
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                target_cents: { type: integer }
                stream_id: { type: string, nullable: true }
                is_active: { type: boolean }
                is_complete: { type: boolean }
      responses:
        '200':
          description: Goal updated
    delete:
      tags: [Moderator]
      summary: Delete a fund goal
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key matching `ADMIN_API_KEY` env var